Switch type of similarity threshold to `int`

+++ src/tree.h
@@ -26,7 +26,7 @@ bk_node *bk_tree_node_new(uint64_t hash, uint64_t file_id);
 /* Search for nodes with hashes with distance less than tolerance */
 void bk_tree_search(bk_node *node,
                     uint64_t hash,
-                    size_t tolerance,
+                    int tolerance,
                     anu_vector *groups_out);
 
 /* Insert hash into tree */
